Add game() function and make changes to existing functions
The game function runs n number of rounds to determine at the end
who is the winner. If a tie occurs, an tiebreaker round is played. The
default value of n is 5.
change is made to the playRound, where the output strings are
modified so that they can be used to easily determine the win or loss
in a round by game(). Moreover, all calls to alert are replaced by
appropriate console methods