Add helper function, modify existing functions, and other changes
The helper function getStrongerSelection() takes either of the
selection and returns a string representing the alternative
that can beat the parameter.
This helper function is then used to modify the playRound function,
where the chain of if statements is replaced by a shorter chain, with
less hard-coded string returns. The code, including the loop, to
resolve a tie is removed and substituted for a return string that
informs of the situation to the caller function.
The literals named after the possible selections were capitalized, and
hence, the getPlayerSelection() function is modified to adapt to This
change.